// src/common/enums/log-event.enum.ts
export enum LogEvent {
  MESSAGE_QUEUED = 'MESSAGE_QUEUED',
  MESSAGE_RECLAIMED = 'MESSAGE_RECLAIMED',
  PROVIDER_ATTEMPT = 'PROVIDER_ATTEMPT',
  PROVIDER_RETRY = 'PROVIDER_RETRY',
  PROVIDER_FAILOVER = 'PROVIDER_FAILOVER',
  MESSAGE_SENT = 'MESSAGE_SENT',
  MESSAGE_FAILED = 'MESSAGE_FAILED',
  MESSAGE_DLQ = 'MESSAGE_DLQ',
  DLQ_PUBLISH_FAILED = 'DLQ_PUBLISH_FAILED',
  MESSAGE_REQUEUE_REQUESTED = 'MESSAGE_REQUEUE_REQUESTED',
  MESSAGE_REQUEUED = 'MESSAGE_REQUEUED',
  REQUEUE_FAILED = 'REQUEUE_FAILED',
  REQUEUE_REJECTED = 'REQUEUE_REJECTED',
  WEBHOOK_RECEIVED = 'WEBHOOK_RECEIVED',
  WEBHOOK_SIGNATURE_INVALID = 'WEBHOOK_SIGNATURE_INVALID',
  WEBHOOK_PAYLOAD_INVALID = 'WEBHOOK_PAYLOAD_INVALID',
  WEBHOOK_DUPLICATE = 'WEBHOOK_DUPLICATE',
  WEBHOOK_MESSAGE_NOT_FOUND = 'WEBHOOK_MESSAGE_NOT_FOUND',
  WEBHOOK_STATUS_UPDATED = 'WEBHOOK_STATUS_UPDATED',
  WEBHOOK_STATUS_IGNORED = 'WEBHOOK_STATUS_IGNORED',
  WEBHOOK_STATUS_CONFLICT = 'WEBHOOK_STATUS_CONFLICT',
  MESSAGE_DELIVERED = 'MESSAGE_DELIVERED',
  IDEMPOTENCY_HIT = 'IDEMPOTENCY_HIT',
  IDEMPOTENCY_LOCK_ACQUIRED = 'IDEMPOTENCY_LOCK_ACQUIRED',
  IDEMPOTENCY_CONFLICT = 'IDEMPOTENCY_CONFLICT',
  QUEUE_PUBLISH_FAILED = 'QUEUE_PUBLISH_FAILED',
  DUPLICATE_JOB_SKIPPED = 'DUPLICATE_JOB_SKIPPED',
}